pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phone-number-backend"
version = "0.1.0"
description = "Monorepo for phone number search backend and admin services"
authors = [{name = "Devanshu"}]
requires-python = ">=3.11"
dependencies = [
    "Django==4.2.11",
    "djangorestframework==3.14.0",
    "drf-spectacular==0.26.5",
    "django-cors-headers==4.3.1",
    "django-ratelimit==4.1.0",
    "python-dotenv==1.0.1",
    "openpyxl==3.1.2",
    "prometheus-client==0.17.1",
    "python-json-logger==2.0.7",
    "psycopg[binary]==3.1.18",
    "gunicorn==21.2.0",
    "uvicorn[standard]==0.24.0",
    "pytest==7.4.4",
    "pytest-django==4.7.0"
]

[project.optional-dependencies]
dev = [
    "ipython",
    "black==23.12.1"
]

[tool.setuptools.packages.find]
where = ["shared", "services"]
include = ["*"]

[tool.pytest.ini_options]
testpaths = [
    "services/api/tests",
    "services/admin/tests"
]
addopts = "-ra"

